+++ ngaq/src/tools/I_StrSegment.cs
@@ -2,5 +2,5 @@ namespace tools;
 public interface I_StrSegment{
 	str text{get;set;}
 	
-	int start{get;set;}
+	u64 start{get;set;}
 }
