+++ ngaq/src/tools/StrSegment.cs
@@ -2,5 +2,5 @@ namespace tools;
 
 public struct StrSegment: I_StrSegment{
 	public str text{get;set;}
-	public i32 start{get;set;}
+	public u64 start{get;set;}
 }
